minimize view preset
ref #33
