+++ src/app/modules/attribution/view-maprule.css
@@ -12,6 +12,7 @@
     background: #f8f9fa;
     margin: auto;
     border-bottom: 1px solid rgba(0,0,0,.125);
+    padding: initial;
 }
 
 .view-label {
@@ -21,6 +22,31 @@
 
 .view-label h5 {
     font-size: 30px;
+}
+
+.preset-accordion-button-container {
+    margin: 10px;
+    margin-top: auto;
+    margin-bottom: auto;
+}
+
+.preset-accordion-button {
+    height: 50px;
+    width: 50px;
+    border-radius: 10px;
+    border: 1px solid rgba(0,0,0,.125);
+}
+
+.preset-accordion-button:hover {
+    border-color: rgba(0,0,0,.2);
+}
+
+.preset-accordion-button:active {
+    transform: translate(0,3px);
+}
+
+.preset-accordion-button i {
+    font-size: 20px;
 }
 
 .view-panel-content {
